Update Category service for security check

+++ Application/Categories/EditCategory.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Core;
+using Application.Interfaces;
 using AutoMapper;
 using Domain;
 using FluentValidation;
@@ -28,16 +29,20 @@ namespace Application.Categories
         public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext context;
+            private readonly IUserAccessor userAccessor;
 
-            public Handler(DataContext context)
+            public Handler(DataContext context, IUserAccessor userAccessor)
             {
                 this.context = context;
+                this.userAccessor = userAccessor;
             }
 
             public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var category = await context.Categories.Include(a => a.Attachments).FirstOrDefaultAsync(a => a.Id == request.Category.Id, cancellationToken: cancellationToken);
                 if (category == null) return null;
+                var me = await context.Users.FindAsync(new object[] { userAccessor.GetUserId() }, cancellationToken);
+                if (category.RestaurantId != me.RestaurantId) return Result<Unit>.Failure("You can't edit categories made by someone else");
                 category.Name = request.Category.Name;
                 context.RemoveRange(category.Attachments);
                 category.Attachments = request.Category.Attachments;
